Refactor expression compilation into utility
